[feat] 작성자 fetch 제거 populate 이용
