+++ src/app/(suwan)/my-study-main/[studyId]/board/task-board/[taskId]/page.tsx
@@ -1,4 +1,4 @@
-import { fetchBlackboard, fetchUser } from '@/utils/my-study-main/fetch';
+import { fetchBlackboard } from '@/utils/my-study-main/fetch';
 import TaskDetail from '../../_component/TaskDetail';
 import { getSession } from '@/utils/getSessions';
 
@@ -14,11 +14,9 @@ export default async function page({
   const taskId = params.taskId;
   const task = await fetchBlackboard('task', taskId);
 
-  const writer = await fetchUser(task.writerId);
-
   return (
     <>
-      <TaskDetail blackboard={task} writer={writer} userId={userId} />
+      <TaskDetail task={task} userId={userId} />
     </>
   );
 }
