+++ src/types/TStudyBoard.ts
@@ -3,3 +3,20 @@ export type TFormData = {
   content: string;
   img?: string | null;
 };
+
+export type TBlackboard = {
+  _id: string;
+  studyId: string;
+  writerId: {
+    _id: string;
+    nickname: string;
+    image: string;
+  };
+  type: string;
+  title?: string;
+  content?: string;
+  image?: string;
+  views?: number;
+  createdAt: string;
+  icons: any;
+};
